Write and test functions for original Olsen-Randerson downscaling.
I would like to extend this to include the Fisher et al. (2016)
modifications to the original Olsen and Randerson (2004) method.

If some model outputs GPP and Reco/Resp instead of just NEP or NEE,
users can downscale the components separately.

I would like to extend this to work with ensembles of NEP for a single
month.